gfxstream: re-add a codegen for gfxstream
This is the same version of codegen as in Mesa 8ef0de25.

A pain point identified by the recent Mesa move is the
inability to update the host codegen scripts and host
codegen output at the same time.

This is because host codgen is now housed in Mesa.  So
one would have to make a Mesa commit and gfxstream commit
to make a change.  Plus, there's the fact there's also
AOSP Mesa too, so the exact state of codegen at a particular
point is harder to reason about.

The long-term best solution maybe an unified repo inside Mesa
again with host too, but that might take forever.

A simple way to mitigate the confusion is just to keep two
copies of the code: one in gfxstream and one in Mesa.  Based
on the observation that host codegen updates are much more
frequent that guest codegen updates, many developers can just
modify the host code as usual.

Many times, the guest side doesn't need to be updated at all
for a particular host codegen change.

In the case a codegen change does a guest side effect, we can
just cp this host codegen variant to the Mesa variant.  History
preserving methods can also be done via git format-patch, hence
the directory structure is similiar "src/gfxstream/codegen/scripts".

But again, with the observation guest codegen is less modified than
host codegen, we can conceivably just update the Mesa codegen scripts
every {week}/{month} (take your pick).

BUG=369906368
TEST=CI

Change-Id: I0a24e3772ca555b411b75343c41f9a04b3b64a67