wip: try to improve snapshot: always save, but selectively load
do not submit, still work in progress

There are many vulkan apps that are not well supported
at the moment for snapshot. And the current way to handle
this limitation is to disable snapshot entirely for the guest
if there is any vulkan app launched by the guest that is not
in the allowlist. This seems too restrictive.

This CL tries to be more flexible. If an app is in the allowlist,
its vulkan state will be saved and loaded, the app will resume
its rendering state; on the other hand, if an app is not in
the allowlist, its state will still be saved, but during
load its state will be loaded and discarded, which might
cause the app to crash in the guest, which seems not the
worst thing that can happen, as the guest still runs,
and the allowed vulkan app still continue their states.

In the future, we might discard the save for the app
that is not in the allowlist, this can save some cpu
cycles.

Bug: 370344466
Change-Id: I02f9eef5a4327b73a31e7d9fd1f9d47d9c67b11a