Revert "Physical GPU selection fixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe36ace3-e803-4adb-9d41-d8e9cfb982da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47315fbb52a8fb07daa9e1513dae1983416beb95
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com