Revert "gfxstream: remove references to Mesa in host build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3db074df-2255-479b-b19e-ab00c86fd1c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44a294dad1837966dc5a37b3f387ebe80b541afb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com