Revert "gfxstream: remove references to Mesa in host build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04333529-1bde-4a2b-b5ae-0d80c149bfd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d6fe277211b334030cad3af11d5c159d79844b6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com