Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/462151a0-5c56-4c6b-83f3-6d3de57767b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabcc6fc1a2187a78559031fc4aff42d04eb8902b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com