Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/462151a0-5c56-4c6b-83f3-6d3de57767b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ide8f5642881a75f27d1410a8b4885bd88c2befae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com