Revert "Update autogen for VK_KHR_line_rasterization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d170ffd-d557-45e6-86d1-1f4f959ec002). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec77f391ea8fa50b1a7ca1d42503ef58adeeba86
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com