Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b78a14c-d4ef-4014-ab01-80e3d224254e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9edb6fe9daf45f3b854e76cb94af9fa2ec8e454b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com