Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b78a14c-d4ef-4014-ab01-80e3d224254e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic04ae2fd8a11d7a5d3bb2c5f2dea7d39f7c93fd1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com