Remove virtio gpu context id from blob keys
Including contexts id in the keys in ExternalObjectManager makes
processing the command stream stateful/dependent on which
RenderThread/VkDecoder processes the command stream which
complicates snapshots. There is not currently a way to inject
additional state into the saved API traces nor is there currently
a way to replay the API traces on the "same" RenderThreads that
originally processed the calls. Slightly more info in
go/gfxstream-snapshot-stateful-replay for more info.

Bug: b/319729462
Bug: b/369615058
Test: cvd start \
      --gpu_mode=gfxstream_guest_angle_host_swiftshader \
      --gpu_renderer_features=VulkanSnapshots:enabled \
      --enable_virtiofs=false
Test: cvd snapshot_take \
      --force \
      --auto_suspend \
      --snapshot_path=/tmp/snapshot1
Test: cvd create \
      --snapshot_path=/tmp/snapshot1
Change-Id: Ifa67382399fb49fd019baf0026ff68cf1c872cec