Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36d3dbae-725c-4817-b9c4-9c16b2fdd5a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf8a0ee868ccac432a3ae7bea0107ba84f852a69
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com