Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3799c6f0-b2bf-4a04-882c-02cb775daf38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I93f5af2d419f587a8783d80c3edd7f49f8553fc2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com