Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/79eb140c-2719-44b5-909e-25022e5d1d61). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib228cf85f80e301009afcd688c6c9d0570546914
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com