Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/79eb140c-2719-44b5-909e-25022e5d1d61). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieacb749f7ebadb849c1d13784fca3ac824ddae05
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com