Revert "Change log level to info for vulkan device selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb4a75fb-4214-4935-b5cf-38ca72485989). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I96bd97c2943ed1789b33e97e2a44593ec49288b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com