Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ba4e23f-50e3-492c-91a6-65aa123a22a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie3a37f424645289ec5fd07d829a74b5a7ca30160
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com