Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ba4e23f-50e3-492c-91a6-65aa123a22a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3aaa2a61f791368c04c420b84ec1a02a49dd2af1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com