Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46ec42f3-013a-403d-9414-621e21e1cdb3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I41772342ebe145d56ac9532ca8b492d82777a2f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com