Revert "Remove unused mResourceContexts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c67fc40-e35e-4d05-a97d-3196512dae82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I969c4de3a66379908ffc3b8a1036814f68591930
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com