Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59c5d002-4e3c-4a47-abca-04a52dd705f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iccbc8a530cfb2e5e5a57ecd9c64e2c6c152ff556
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com