Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbf6baa-c77d-4484-b5c3-89b5d21b26cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8c0172032cc0fdf7753e11130feb31c84121809
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com