Revert "Change log level to info for vulkan device selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/733f0b28-f9ab-4912-88a5-1f0d2a073a43). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I707bcddb88f1a85f89a1500bd20e9645c6428369
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com