Fix default behavior of vulkanBatchedDescriptorSets from stream-renderer interface
This recent change: https://android-review.googlesource.com/c/platform/hardware/google/gfxstream/+/3323076  is not backwards-compatible. It breaks existing Android guests, which have no option to disable vulkanBatchedDescriptorSetUpdates from the guest. Also, the feature_condition added as part of this change does not make sense (just sets feature.enables = feature.enabled).

The default (at least when initialized with the stream-renderer API) must remain ON. Introduce a new flag to disable this. In the future, this might converge to always be disabled.

Change-Id: I4c542d74301474f640da03d2d67a0586c9b0f416