[Vulkan Snapshot]: handle vkCmd apis
Currently, DeferredVulkanCommands is used, which
stuff all the vkCmd apis into the SubDecode, and treat
it as a whole api.

The problem with that approach is the lack of dependency
handling of cmd on resources. For example,
vkCmdPipelineBarrier should depend on the image or buffer
but it is not clearly expressed in current dependency
graph.

so the plan is to disable DeferredVulkanCommands when
snapshot is turned on so we can handle dependency
of cmd on its resources explicitly.

The steps are as follows, star from vkBeginCommandBuffer,
create a generic action handle fo each command,
make vkBeginCommandBuffer action depdends on <commandbuffer-Handle, CMD_RECORD>
make vkCmdx action handle depends on vkBeginCommandBuffer action Handle
make vkCmdy action handle depends on vkCmdx action Handle
..
make vkEndCommandBuffer action handle depends on last command action Handle

note, each <commandbuffer-handle, CMD_RECORD> should have
at most one child.

However, one possible problems:
the vkBeginCommandBuffer ... vkEndCommandBuffer
should go as a set, and there should not be
any middle command mixed. how to achieve that ?

actually, this is not a problem, since each
command has its first parameter of cmdBuffer,
so they wont mix up.

the ordering is already embedded in the dep
graph, so no need to have the separate
modify apis.

Bug: 370344466
Change-Id: I74b09d85f6e240564e93713183bfcf68e1671dad