Delay the cleaning of finished render channels
Previously, the finished render channels are
closed quite early during the create new
render channel function call, which is often
invoked by the vcpu thread holding the big
qemu lock. It has the benefit of timely cleaning
resources, but has the draw back of causing
hang detector to kill emulator if the cleaning
process is blocked wating for the rendering
thread to join.

So to avoid hanging, this cleaning of finished
renderer channels is moved to the end when the
whole rendering app is done. This might temporarily
increase resource usage, but seems a good trade
off for better stability (fewer hanging).

Test: run VK.api.* test and emulator should
survive without being killed by hang detector.

Bug: 379737524

Change-Id: I468ee17f79d916acad51e20a2e117846e8a81b70