Revert "Rename "pipe renderer" objs to "virtio gpu" objs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7abbd0c3-f76f-42ea-be1c-0c1a73a3fe24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d08fb400af58889492803c84d22a0ba2433c898
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com