Revert "Add suspend/resume stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7abbd0c3-f76f-42ea-be1c-0c1a73a3fe24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic78cfe5462a8208992f6c0f572d88028eb990406
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com