Revert "use finer lock in FrameBuffer for ProcessResource"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f6575ac-c3af-4570-a1ba-532be552dafd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1b979b19bca19ac1decc048012e10c3cd70b056
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com