Enable VulkanDebugUtils by default
Remove the feature flag for enabling vulkan debug utils, as it's not
working properly on goldfish and requires extra changes. The overhead
of command and resource markers is expected to be very low, and they are
very useful for debugging. ANDROID_EMU_VK_NO_DEBUGUTILS=1 can be used
to disable in case of any regressions.

Bug: N/A
Test: run emulator with Vulkan API dump layer enabled
Change-Id: I0966311da30187e1ff07c24a90a894d507cdc3c0