Revert "E2E Gfxstream test for Double Queue Gfxstream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/198b8df2-b204-47cc-87d1-1d8d7810c919). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I84ce9c7b9dddd7fdc2b496b053c22b008becb7d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com