Add new interface stream_renderer_import_resource
- Add ResourceDescriptorInfo map to ExternalObjectManager, to hold the
ExternalHandleInfo for fro resource_create_import_handle
- For Vk ColorBuffers, pull the external memory handle for import from
the ExternalObjectManager

Evidently, this is very similar to the stream_renderer_platform_import_resource interface.
This interface will be replaced by this new one in a separate CL (coming
soon). When this happens, create_args will be optional, indicating that
we are importing to an existing resource.

BUG=382442209

Change-Id: I8eeefd75fe0593b659327b20af51dfa84d5b7ca9