Revert "Enable mTextureDraw deletion"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a04c944-de70-43b6-ade3-bcfac95a618b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383863009
Change-Id: I8d74569bbf005143b082593fc420936f75b8d4e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com