Revert "Report proper buildid for usage in crashpad"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7a152fe-b3b8-431c-b80a-231bac214df6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8040719cbfcde146a8ddca5887b787be41e2a726
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com