Revert "gfxstream: fix meson build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2019f7d7-02ed-4f8e-98cd-038b36c579a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5184b588b17ce839095155b4acfe2b9a91a2421
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com