Track pipeline layouts on decoder
Track pipeline layout creation and destroy calls
to cleanup them correctly on device teardown.

Pipeline layouts require delayed delete operations for
VulkanQueueSubmitWithCommands feature which modifies order
of commands. Pipeline layouts needs to stay valid until the
submission time. With this change, we move that logic to the
host side to have better control based on the enablement of
of the feature flag and delete them immediately when possible.

Bug: 342343008
Test: run and shutdown gltfviewer with VVL
Change-Id: I465a8e3f6ed7e621470f32cdb7cda91ec3dcbba6