Revert "serialize syncThread creation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36e7982b-4000-4440-82e8-195eef2d4667). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4bcd1673b9f6894ab59fcea9528ec88780ccca4c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com