Better handling of fatal Vulkan errors
Introduce onFatalError api to handle some of the fatal
errors on the host side when dealing with vulkan. This
would mark the render thread and related devices as lost
which will then cause dispatch_Vk* calls to return null
for the device. Additional dispatcher checks are added
into codegen to handle invalid cases, later we can also
replace the dispatcher object with a dummy one to avoid
doing so many null checks.
These changes are aiming to get snapshot system working
by marking the devices as lost on snapshot load and let
the applications handle errors, and regerate their vulkan
objects themselves.

Bug: 383272733
Test: run buggy vulkan apps, emulator should not crash
Test: enable vulkan snapshots manually, save&load should only crash guest apps
Change-Id: I3ace0eee370f1694ce74db49bf0c1980324eea7b