Revert "Unregister cleanup callbacks before destroying instances"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f21ba21-fe5c-4d14-990a-953be2081ff0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I146d77322798c6fa3241b44832a3f5331b57889b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com