Send host the name of rendering app
process_name usually corresponds to the app
name, but not always.
for example, deqp app sometimes append ":testcores"
to the app name as process name, and in that
case host side needs to strip it off.

the purpose of this is to let host side selectively
force stop the app, if it ever losts the device.

Bug: 390047716

Change-Id: If42986f0354cfa99759fccdb8325fb4894dcef6c