Revert "Add missing fence reset for commandbuffer test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74d2055c-80e4-4cf1-845b-72bc95425153). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1303c051da04995ea34faf85f02b1accda6f8e57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com