Revert "snapshot: register vulakn apps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78baee7e-74f3-4aa4-8c7e-9f64d1df4056). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48d61b8d44ad16c52e7a809fff902482c0e71cc9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com