make vkdecoder global state process local
do not submit

problem: syncthread gets a state that is isolated
from process' and it does not seem to be correct.

apparently, some ids are used cross proccesses, such as
Image/fence, and for them, we might have to have some
kind of global data structure to hold them.

vast majority of other ids, are local to process ,and
we should make it that way.

Change-Id: I2a588e4046da2f70babf0096ce2408e0696c5d13