Revert "Add missing fence reset for commandbuffer test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de9c847b-2bef-4894-bd5c-4839f4598958). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6773238c694819657897ea7e0fda8d068552aeb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com