Do not check fence state when not waiting for all fences
We started waiting for the correct fence state with recent
changes, but that doesn't support the case where only some of
the fences were signalled. We limit the checks for the state
only to cases where all fences are known to be signalled.
This should not change the old code path behavior for direct
calls to waitForFence(), which is always called with a single
fence.

Bug: 397501277
Test: CI && run apps on TV api level 31
Change-Id: I349b51a3f9c7929802a9c1d62ea96cf5df50cc6a