Revert "Sending process_name properly using rcSetProcessMetadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8c62d67-11c9-408a-ac14-93f64641e4e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75ba47faf811775653c61a43212aaa9c432f679c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com