Revert "Add missing fence reset for commandbuffer test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/414d2d8d-6078-4791-9e4e-9722df86a474). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33163fbb59b849abe5b816c07b578603f9c557ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com