Revert "Sending process_name properly using rcSetProcessMetadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/432e0f24-d599-46d7-b1f3-33f3e71d1be7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I874290280289e6016869aff604d8ffbd507b2576
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com