Revert "Revert "Revert "gfxstream: stream-renderer: add waitSyncResource...""
This reverts commit 1cc4e5cfc827c201545c09de4d747d3ed36cbbae.

This is now a "revert^3". So finally removing this API/feature,
and all the associated code that was tracking this in queueSubmit.
As mentioned on chromium review, this was initially added by QNX for
zink+gfxstream-vulkan bring-up on Linux, but sync issues fixed in
Vulkan. This was never a robust API, and doesn't make sense in the
context of the virtio-gpu spec (i.e. for scanout resources).

Test:Linux + Android guests using gfxstream-vulkan-only configs

Change-Id: I9e25f797b27388f78c19b2fb6d979fd803bce60e