Revert "Sending process_name properly using rcSetProcessMetadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18baedbe-912f-4da1-834a-702ede70164e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4d8b05b1f9eff2d338f24ad8a60583ac1164c63
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com