Revert "Enabling lavapipe --gpu lavapipe option"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a185257-56fa-46f8-97e9-19c2d10e5fd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf4187a00fd775504f95c9a411bb285fc206d5c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com