Revert "Disable -Wcast-function-type-mismatch"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0782d58-382f-44b6-8cd8-b2e4ba15fef8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic188dff5d8eadd3e6f87d991f637120b770a89b2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com