Revert "Disable -Wcast-function-type-mismatch"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4114aa15-0cc6-4776-b4ff-a6fa902ae5d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38105bec0579ef936e9978e1fb750c2033be213b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com