Handle timeline semaphores with virtual queue
Special handling of submissions where the signalling will be done
after a wait operation on the queue.

When a single physical queue is shared with VulkanVirtualQueue, any
subsequent signals cannot be processed as the wait operation would
block the queue. We defer the real submission until another queue
submission with the necessary semaphore signaling is made. For other
post-submit operations, we treat the pending ops as if they have been
submitted, because the object lifetimes management (e.g. semaphores,
fences, command buffers) should be controlled by the app for valid
usage and the fence for the host operations won't be signalled until
the sinal-execution chain is completed.

We don't support submissions with timeline semaphores on presentation
images, so copies to gles backend with explicit fence waits should
still work fine.

Bug: 398986781
Test: dEQP-VK.synchronization2.timeline_semaphore.wait_before_signal.*
Change-Id: I411ff1524b835658b055083acafb162bd9e5546e