Try to use the current dynamic library search path to search for ICD files.
This is necessary as, in the new emulator, the ProgramDirectory and LauncherDirectory are both under external/qemu. However, the launcher is actually under hardware/generic/goldfish and that's where the libraries are gathered. DYLD_LIBRARY_PATH is already set by the launcher so that Qemu can find its plugins and gfxstream is also able to find libvulkan etc that way.
Given the relevant path is already being communicated through this env var, it seems simplest to use it to find the icd file too for now.

Bug: 393139139
Change-Id: I821bc6c5ce031353e7668f7eed90c65791ffa084