Better handling of fatal Vulkan errors
Introduce onFatalError api to handle some of the fatal
errors on the host side when dealing with vulkan. This
would mark the render thread and related devices as lost
which will then cause dispatch_Vk* calls to return null
for the device. Additional dispatcher checks are added
into codegen to handle invalid cases, later we can also
replace the dispatcher object with a simpler one to avoid
doing so many null checks, and return error with codegen.

These changes are aiming to improve stability of the emulator
and also get snapshot system working by marking the devices
as lost on snapshot load stage. This will let let apps to
handle the errors, and regerate their vulkan states directly
on the guest side when necessary.

Bug: 383272733
Test: run buggy vulkan apps, emulator should not crash
Test: enable vulkan snapshots manually, save&load should only crash guest apps
Change-Id: I1ee536da065ecba157e6de068040fef12463ca14