Revert "Address a -Wshadow error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd62cf01-3255-4dc9-9d7d-a46e2bef6a27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401014524
Change-Id: I1b6486133261cfb31acc33e7c8613eb855ef53c6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com