Revert "Enable -Wshadow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd62cf01-3255-4dc9-9d7d-a46e2bef6a27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401014524
Change-Id: Ib0a6577d014285b6005d954bebbfe5cccf385840
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com