Revert "Enable -Wshadow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5555cb36-a8da-493f-907e-cd83e9ca4bcf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401015047
Change-Id: I4715f9bcad9aac0942a83d3c37fa5f9d0d24e69d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com