Revert "Address a -Wshadow error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5555cb36-a8da-493f-907e-cd83e9ca4bcf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401015047
Change-Id: I50cfdff12aea46c4311354b0af7c2274a0b756a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com