Revert "Move vkDispatch to after emuglConfig_setupEnv so that it..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcb7c920-68a4-4dbf-b5b6-a4ae09f3da90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I577e07e956ae2870f14524ebdeacab3ed4323f5f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com