Revert "Move vkDispatch to after emuglConfig_setupEnv so that it..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08bf9565-5b9d-4964-8cab-94e175a27806). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26fbd9dd145f653212581b41f38d26610ef4e132
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com