Revert "Adjust level of some log messages"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c7fb9a42-4dbe-42b9-86dc-f0d146885022). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1833d9c7a0629fde9e00a9e45158e61199683ddc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com