Revert "Workaround the issue in b/399729237 instead of reverting..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c992233f-110b-47ea-abb6-2d9c7eec4c92). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c62b0e16f43487281f4916880ed9150ac979c36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com