Revert "Serialize vkEnumerateInstanceExtensionProperties on linux"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e9c06f9-5c9d-40ae-b002-f417cf2127d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2dc76ed6d85c924fefa30db0ae99609cf486322f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com