[vulkan-snapshot]: assign VkSnapshotApiCallInfo correctly
There is a race condition that can invalidate
the VkSnapshotApiCallInfo. so use the id to
retrieve the up-to-date VkSnapshotApiCallInfo
and update there.

e.g.
thread A created VkSnapshotApiCallInfo, obtained a pointer
to it;

thread B created another VkSnapshotApiCallInfo, invalidated
the previous pointer, as there is a resize in EntityManager.

thread A updates its invalidate pointer's content, but it
has no effect on the updated VkSnapshotApiCallInfo with
the same handle, which leads to the updated VkSnapshotApiCallInfo
with empty packet, as the assign to the packet is performed
on the invalidated VkSnapshotApiCallInfo.

the root cause is that the lock is not applied throughout
the whole decoding process, giving other thread a chance
to grab the lock and invalidate the VkSnapshotApiCallInfo
before ThreadA is done with it.

Change-Id: I51ee441ea445b1049a08efa7e7002a5a12cfd306