Revert "Reduce log level on protected command pool flag modification"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/157acff7-eedf-4bfc-af7a-76b75c2291b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3be696341a65a3f10ef53a18237ac5a699257edc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com