Revert "Re-enable -Wshadow and fix remaining issue"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f53b5d52-e9ef-45fc-8b8d-0bab269e6b38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic1d1d8c1464f47e577b3c05604e4646abf4e8140
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com