Revert "avoid using un-initialized queue family index in VkAndro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fa97018-30ed-410f-b4f6-20a780b00ec0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea5a45f18efae24f93b39e7eb67d83561d6849a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com