Revert "Remove FeatureDependencyHandler"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cae02b59-290b-4b15-bdf1-00ed58cc8bff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf4b17eadbbeb60e376de41a53b7092a6f9190f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com