Revert "BoxedHandleManager: fix corner case where there is no re..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/93b6c4f5-d71c-4dff-b468-fb24ce2736da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6dfd565740ec56f2c944946fccbeb21793e55ab7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com